Fix bug in index checking.
This came through a bug where `!paired?` incorrectly rejects all reads using
the index. Some predicates do not check the index at all. If they pass through
not or exclusive-or, then they incorrectly get dropped.

This introduced a secondary check to decide if a predicate will bother to check
the index and generate a more sensible checker.